data_collection/requirements.txt
langchain
langgraph
langchain-core
langchain-tavily
langchain-openai
pydantic
pandas
python-dotenv